Fixed Various Issues
Fixed Admin prescription approval, imlemented product searching, Improved Products UI , Fixed Order amount inconsistency, fixed customer prpfile updation,  fixed phone field for executives, Added Help, Improved DoctorPreview

+++ Backend/Curevo/src/main/java/com/certaint/curevo/service/CustomerService.java
@@ -117,7 +117,9 @@ public class CustomerService {
             }
 
             if(image != null && !image.isEmpty()) {
+                if (existingCustomer.getImage()!= null ) {
                     imageHostingService.deleteImage(existingCustomer.getImage());
+                }
                     String imageUrl = imageHostingService.uploadImage(image,"customers");
                     existingCustomer.setImage(imageUrl);
             }
